Rework of channel join logic upon connect
It used to be that the trigger to join the channels listed in the
configuration was NickServ's acceptance line, which caused problems
if the requested nick was taken. There would be no challenge to
respond to by identifying, and as such no acceptance message to
join channels upon receiving. Changed this to trigger on end of
MOTD instead.