Proactively append the altNickSign if nick taken when logging in
The original logic was that only the request to change to the new
nick should be sent, and the IrcBots only updated upon receiving
WELCOME, but this caused an endless loop when a nick was actually
taken. It would request nickname ~ altNickSign, get a refusal,
re-request nickname ~ altNickSign, until stopped.