Add new variadic templatesend-line-to-server logic
It used to be that the strings had to be manually formatted and/or
concatenated if you wanted a more complex string sent. Now we instead
define a variadic template that foreaches over all arguments and
discretely sends them in turn. This might have posed a race condition
depending on how Socket.send works, but only one thread is ever sending
things, so it doesn't apply.