Move state structs to module-level
There's really no point to having IrcPluginStates be members of the
plugin classes. We enjoy thread-local storage so it's perfectly
thread-safe to keep a top-level IrcPluginState for the whole plugin
to use. That way we don't need to pass it around as ref to every
other function.

If it ends up being modified in too many places we can just create
mutators to consolidate things.