Add kameloso.stringutils.timeSince
It takes a Duration and (currently) splits it up into days, hours, minutes
and seconds if less than a minute. It's needed for the coming changes to
inverting the Notes plugin's reporting of time since a note to a user was
saved.

Current behaviour is to send at what timestamp the note was created, but
wanted behaviour is like that of feepbot; sending a humanly-readable
description of the duration since then.