Catch the base Exception instead of specific ConvExceptions etc
There's no need to disambiguate if we're not treating them differently
from each other. We're only polluting the available symbols.

As a notable exception SocketException stays as is, but only because we
know those are benign and we don't know what kind of other Exceptions we
might run into that aren't.