Disable the pinger
Initially the Pinger was created to proactively ping the server, as it
"just happened" to stop.

It turns out if you pong the first ping with the correct reply, as
the TOCONNECTTYPE event suggests, it sets up a proper ping/pong cycle.

Now that we know this, disable the Pinger plugin, and remove it later
when we're certain it's not needed.