Remove Decode.(yes|no) Flag on stringutils.nom
Up until now we've kept a compile-time flag on nom to specify whether
it should use std.string.indexOf or std.algorithm.searching.countUntil
internally, as the notion was that indexOf decodes while countUntil
does not.

Simple benchmarks show that countUntil is upwards of 7x slower, so
remove the flag entirely and always use indexOf.