Fix default-case bug in userFromEvent
It was looking for the character '@' in the event sender string to tell
if the origin was a user or the server, likely to tell if it was a
string of the kind 'username!~ident@address'. However, at that point
we have already split that into nickname, ident and address.

Look for a non-empty ident instead.