Use countUntil in nom if template parameter set
std.string.indexOf is much faster than std.algorithm.searching
.countUntil on strings, but countUntil wins out ever so slightly on
arrays of ubytes. We know part of server strings are directly from
the server, so at least those we should be able to treat directly
as ubyte[]s and enjoy the speed boost.

The template parameter Yes.ubytes has to be set for this functionality
to kick into gear.