Add mIRC colouring
This translates mIRC colouring of incoming text and translsates it to
the Bash equavilents. It is not a 1:1 mirror with regards to the
colours, but rather a copy of how WeeChat displays them.

One notable missing feature is the italics. Unsure of how to display
those on a terminal. WeeChat manages it.