Add meldInto
It takes two structs and assigns values to struct members from the
first to the second if the second's members are unset (.init).
With this we can read configuration from disk while at the same time
gathering settings from the command line, and then apply the disk-
read version onto the command-line-read version, filling in the blanks.

This fixes the issue of the -c argument (for specifying config file)
not working with manually specifying arguments.