Fix LDC FTBFS by making some of irc @trusted
What's happening is that in the normal dmd build we're passing
arguments to formattedRead by ref, which LDC doesn't seem to manage
at its current frontend version. If we explictly pass them as
pointers however, we violate @safe by taking the address of locals
at certain points in the code.

Annotate those functions as @trusted for now to make it compile under
LDC. Revisit later, optionally file a bug report.