Add tentative YouTube webtitles workaround
If your IP has been flagged as spamming, YouTube will refuse to serve
video pages, instead claiming 402 and/or requiring you to solve a
Captcha.

Since we can't very well do this from the terminal, rewrite the
youtube.com URL into a youtubeonrepeat.com one. It doesn't seems to
trigger the same behaviour.

We'll just revert this if it doesn't work.