Rewrite sedReplace to work with dmd git
The old approach was to have a nested function return which regex engine
to use based on the expression string (s/a/b/, s|a|b|, s#a#b#, etc), and
then do regex matching with it, but this suddenly stopped working with dmd git.

  Error: mismatched function return type inference of Wrapper and Wrapper

alphaglosined in #d on freenode said the error was to be expected, so
rewrite the function to instead have the regex matching happen in the
nested function and the engine selection in the base function.