Add mapAlternatingEffects
It seems that servers use the mIRC colour codes for effects (bold,
italics, underlined) in an alternating pattern; one toke activates it,
then it stays active until the next token comes to deactivate it. We
were only ever enabling the effects (by naïvely doing regex replace of
the token ubytes with the corresponding Bash token strings).

Now we look at them in pairs and in a loop replace them with an opening
token and a resetting token.

So far it uses the catch-all reset token which *will* invalidate
colours, but this should be fixable down the line.