Change IRC event parsing functions to take ref Bot
This way we don't rely on the irc.d module keeping any globals. Passing
it by ref means we can easily flag it with a new bool IRCBot.updated,
signifying that it was altered somehow. Then we can pick up the change
at the top of the stack at the original toIRCEvent(string,IRCBot) call
and propagate it to all plugins.

Neat and tidy, no hook hacks needed.