Initial work on IRCUsers as IRCEvent sender/target
An IRCEvent now has two IRCUsers; sender and target. They in turn have all
the fields previously pertaining to the sender and the target in the root
IRCEvent, which removes some of the clutter. Remaining is rolestring and
colour, and maybe IRCUsers should have an aux as well; we'll see.

This way we're not just limited to one string and have to compete with
finding places to put information. To put into perspective, aux2 was
considered.