Support both server pings and connection pings
During connection to some servers you're asked to ping a certain value,
likely a timestamp, whereas with normal pings you want to ping the
address of the originator. The one belongs in event.contents, the other
in event.sender.address.

This used to work but now we broke everything with the IRCUsers.