Add onJoinMixin to plugins/common.d
It will be mixed in by all plugins with the BasicEventHandlers template.

With the capability extended-join enabled, the login name of whoever's
joining a name will be included in the server string. So we can catch it
then and skip having to WHOIS later.

If the login field is empty, that either means extended-join is disabled
or the user was simply not logged in, so skip those events.