Update bot via passive yieldBot, not concurrency
Upon an updated bot we used to cast it to shared and concurrency-send
it to the main thread loop, but aside from being inefficient it also
ended up thing not playing out in the expected order. So now the plugin
loop will try plugin.yieldBot() before and after onEvent, then propagate
any changes to all plugins before continuing.

Fixes Network deduction not sticking.