New WHOIS and queueing logic
It used to be we sent concurrency messages to the main thread to start a
WHOIS call, who then kept track of when calls were sent so it wouldn't
be too spammy. Once the WHOISLOGIN (or HASTHISNICK) event came back, it
would replay the event for all plugins.

The drawback of this is that, for plugins that would be triggered by the
normal event despite the user's login not being known, the event would
be repeated again. You'd get two calls.

Now each plugin keeps an associative array of WHOIS request structs,
which the main thread will look at and act upon after a plugin has
returned from their onEvent. We just put them there and it'll pick them
up.