Fix SELFNICK not sticking, unit test it
At first we were setting event.content with the target nick, which was
all well and good. After we added target and sender as IRCUser we set
the target.nickname field to point to the same. Then we removed the
content and changed it to sender.nickname by mistake. So all in all, a
mess.

Fix it by setting it to only be target.nickname, and unit test the
change.