Add SELFQUERY and SELFCHAN events
"When a server sends a PRIVMSG/NOTICE to someone else on behalf of a client
connected to it – common when multiple clients are connected to a bouncer –
it is called a self-message. With the echo-message capability, they are also
sent in reply to every PRIVMSG/NOTICE a client sends. These are represented
by a protocol message looking like this:

:yournick!~foo@example.com PRIVMSG someone_else :Hello world!

They should be put in someone_else's query and displayed as though they were
sent by the connected client themselves. This page displays which clients
properly parse and display this type of echo'd PRIVMSG/NOTICE."

http://defs.ircdocs.horse/info/selfmessages.html