Infer changed bot by newBot != bot, not bot.updated
Earlier behaviour was to keep a bool updated that we set whenever we
modified the bot, and then had the main loop pick up on that and
propagate the new one to all plugins.

It's not *as* cheap but it's still in the order of microseconds to
compare the new bot with the old instead, saving us the mental overhead
of having to tack bot.updated = true whereever something happens.