Add message-throttling ThreadMessage.Throttleline
It works like ThreadMessage.Sendline except it throttles outgoing
messages

It allows for an unthrottled burst at first and then works like a
y = k*x + m graph, allowing messages to be sent when the graph dips
beneath a given value (the burst value, default 3.0), sleeping
inbetween.

It's not perfect but it'll do for now until we figure out something
better.