Try-catch each onEvent
This allows each plugin to throw without interrupting the onEvent-chain.
Previous to this one even throwing would still be caught, but no more
plugins would have a chance to onEvent on that event.

Revert if it becomes a problem. Alternatively, add some logic to halt if
three plugins in a row throw.