Don't re-read config into a !T.init setting struct
Now it will only read configuration from disk into structs that have not
been modified; ergo, only once at program start.

It's buggy in that it will still read configuration into structs that
are T.init in the sense that they have default values in all fields. But
we'll live with it.

Fixes #31.