Bake IRCPluginState into IRCPluginBasics
This allows us to remove module-level (global) IRCPluginStates, instead
keeping a copy inside the IRCPlugin object and passing it as a parameter to
functions that want it.

Extend the list of allowed function signatures in onEvent to reflect
this. Update WHOIS logic so it keeps replaying well.

Also allow the IRCPlugins themselves to be parameters. This allows us to
move other globals into its namespace.