Save seen users once every n hours (not hard 24h)
Saving turned out to be prohibitively expensive when in many or very
large channels; or rather, pretty-formatting the JSON for printing was.

We don't need to save that often, we automatically save upon program
exit anyway. But it's nice to be redundant, so save every once in a
while.

Default 6h but configurable.