Add timed Fibers
This introduces an associative array much like that of awaitingFibers,
wherein array of Fibers are keyed by a long UNIX timestamp. This gives
the timed Fibers a resolution of 1 seconds.

Known bug is that the fibers are only processed once an event has been
parsed, making it currently viable in channels where lots of messages are
being sent, but useles in a situation where nothing happens. This is
because of the blocking Socket.