Move Client client into main
This removes the last module-level state variables in main.

The sole exception is the signal handler __gshared bool abort, but it both
needs and should remain top-level. Needs because it has to be accessible
from the extern(C) void signalHandler(int), should because it should be
visible from every instance of Client.

Untested, but it should be possible to run several clients in parallel
now, in multiple threads. Granted, this would wreak havoc with the
not-completely-locking lockingTextWriter used in KamelosoLogger.