Fix timed Fibers only triggering on incoming events
It wasn't so much that the Socket was blocking reading, but rather that
the main loop was breaking and retrying to read if nothing was received,
before getting to the point where it would process the Fibers.

Go through each plugin's timed Fibers even before checking the length of
the received line.