Raise receive timeout from 1 second to 2
This is a test to see if it makes itself felt at all.

The time wasted in the hot path of mainLoop iterating plugins looking
for non-existent Fibers is keenly dependant on how often the function is
called, to no ones' surprise. In the ideal world the function would only
need to be called once per server read and never otherwise, but now that
we both have plugin that run in separate threads as well as timed Fibers
waiting for execution, we can't afford the luxury.

The earlier setting was to time out after 1 second of no text from the
server. Double it.