Make channelqueries Fiber dynamically query aModes
Instead of hardcoding channel A-modes to query, iterate IRCServer.aModes
and query each listed there instead.

Don't rely on awaiting the "return" events (RPL_ENDOFBANLIST and
friends) since we don't know which ones to listen for. Instead, use
delayed Fibers for the MODE queries, and type-awaiting for RPL_ENDOFWHO
and RPL_CHANNELMODEIS.