Bump version, first 1.0.0 beta
Channel awareness is done, needs testing.

User awareness works.

Changing settings by name (./kameloso --set printer.bellOnMention=false)
now works. See ./kameloso --settings or the config file for a list of
all settings.

Available channel modes are now caught during connect and modes properly
tracked as they are added and removed. We're not hardcoded to certain modes
but rather obey what is listed in ISUPPORT CHANMODES.

Prefixing channel/user modes (operator, voiced etc) likewise caught from
ISUPPORT and tracked. Non-channel server user modes are not (+ix etc);
not a lot of work but unsure if there's any point in it.

Fibers can be registered, to trigger at a point in time or be set to await
a certain IRCEvent.Type. Having the timed triggers be there at all is
expensive though, so if they're not used we might remove them and be
lighter.