onEvent: Awareness functions before plugin ones
Execution is in order of declaration, which can cause problems if a
plugin relies on the bookkeeping an Awarenses mixin provides when it
gets mixed in late in the module. Concrete example is seen.d assuming
the existence of a channel in plugin.state.channels, where such would be
created in onChannelAwarenessSelfjoinMixin. The mixin is declared after
the plugin's function is, and a range error occurs.

Sort the functions into those annotated with the AwarenessMixin UDA and
those without, then sort them in that order. It's still in order of
declaration otherwise.