Add UDA Terminating
The inverse of Chainable it denotes that an event-handling function
should be the last function in that module to process if it is
triggered. It is not strictly neccessary since no Chainable implicitly
implies Terminating, but we add it anyway to silence the
wildcard-event-without-Chainable warnings, and to have the code be more
self-documenting.