Only use Regex in BotRegex
The linker issues we were running into were due to the fact that
StaticRegex, which we were overloading the normal Regex constructors
with, is itself an alias to Regex. A better error would have been that
both constructor overloads took the same arguments.

Consolidate rtExpr and ctExpr into just engine, and treat the two cases
(runtime and compile-time expressions) as one and the same.