Add plugins/persistence.d
The Persistence plugin melds new IRCUsers (from postprocessing new
IRCEvents) with old records of themselves.

Sometimes the only bit of information about a sender (or a target)
embedded in an IRCEvent may be his/her nickname, even though the event
before detailed everything, even including their account name. With this
plugin we aim to complete such IRCUser entries with the union of
everything we know from previous events.

It only needs part of UserAwareness for minimal bookkeeping, not the
full package, so we only copy/paste the relevant bits to stay slim.