Rewrite kameloso.bash.invert
The old version used regular expressions to find and replace all
occurences of something (nicknames), but it is prohibitively difficult
to craft an expression that doesn't give false positives.

An expression can be written to match a string surrounded by word
boundaries, but some of the characters valid in nicknames trigger that.

This also adds bool isValidNicknameCharacter(char) as a helper.

Use our new state machine instead. It passes all the tests so it should
work as a drop-in replacement.