Give LogLineBuffers lives, to know when to remove
Without any form of garbage collection the internal list of log line
buffers will grow unbounded. To always remove empty ones during next
PING cycle is a bit much, so we set a lifetime of 20 "lives".

Each PING cycle decreases this figure by 1. When it reaches 0 the buffer
is removed. The amount of lives can be set arbitrarily high as it's only
supposed to solve buffers accumulating *over time*.