Break out trigger auth into MinimalAuthentication
UserAwareness included triggering of functions based on their privilege
level (user class), but it was often the case that functions only needed
that basic function-triggering and not the whole of UserAwareness, which
includes populating and manging the state.users array.

Break out the authenticating-triggering bits into a
MinimalAuthentication mixin, and have UserAwarenses mix that in if not
already present.