Compensate formatObject width/typewidth
It used to be that typewidths eschewed the name widths, by having both
simply blindly trust longest{,Unconfigurable}MemberTypeName and
longest{,Unconfigurable}MemberName. The name width was getting
effectively increased by the typewidth.

Set a minimum typewidth (currently 9, defined as enum minimumTypeWidth)
and use std.algorithm.searching.max to get a width with a bottom cap of
9.

Calculate the name width by subtracting the typewidth and adding the
minimum width, so as to always align with other objects printed with the
same width-by-argument.