Support Base64-encoded authPassword strings
This is merely for a level of security by obscurity. The decoding is
trivially reversed, but it means someone looking over your shoulder
while you cat the .conf file can't see the password at a glance.

If the authPassword string begins with "base64:", decode it when
authenticating, otherwise send it as is.