Fix getNextMidnight not spilling into new months, years
We were using SysTime.roll!"days"(1) to increase the current day, which
along with setting the time to 00:00 made it the next midnight. However,
roll doesn't increase any larger units, so rolling one day from the last
of a month ends up at the first of the same month.

Catch that and use SysTime.add!"months" and !"years" to properly wrap
into the next month/year.