Work around getopt melding false default values
There's no way to tell apart an unset bool froma false one, so melding
fails in some cases where we're melding a default false value onto a
true one. It will assume the melding value is merely unset and skip the
modification.

Work around the issue when the value to set seems to be false.