Don't requeue channel-querying Fiber each iteration
This fixes a bug where the querying function was queueing itself as a Fiber
awaiting a type, accumulatively once per iteration. The correct
behaviour is to only append it once and let it lie until the main thread
sees it's finished executing.

We were being kicked with Max SendQ errors, and we spammed channels.