webtitles.d: Add findURLs
Function that replaces the use of regex for extracting URLs from a
string line.

It uses much less memory and seems to be accurate *enough*.
dlang-requests can always refuse to or fail to fetch the page later up
the stack.

To consider is whether this needs to work on a dstring or if a normal
string is enough. Again, false positives aren't a huge deal, as
dlang-requests will just fail.