Move WHOIS account catching into MinimalAuthentication
It used to be part of general early awareness, but that's too late. A
call to admin's verb "user" on the calling nick will process the request
and print the user's information once the RPL_WHOISACCOUNT or
RPL_WHOISREGNICK arrives, before the actual information is caught and
stored.

Move it into MinimalAuthentication, before request processing.