Add account-querying to onCommandAddAutomode
When adding someone, it will first treat it as a nickname and look it up
in the plugin.state.users associative array. If an entry exists and
there is an account in it, add that account.

If the user doesn't exist or there is no account on record, query that
user with a WHOIS. Create a CarryingFiber!IRCEvent and append it to the
plugin.state.awaitingFibers AAs.

Once the WHOIS reply returns, the Fiber will be run and the user added
iff an account was found.