Add template mixin WHOISFiberDelegate
Mixes in delegates to ease and unify requesting WHOIS results on a given
nickname.

---

void onSuccess(const IRCEvent successEvent) { /* ... */ }
void onFailure(const IRCUser failureUser) { /* ... */ }

mixin WHOISFiberDelegate!(onSuccess, onFailure);

enqueueAndWHOIS(specifiedNickname);

---

onSuccess will be called if an account was found, while onFailure will
be called if the user was not logged onto an account, or if offline.
Both functions can be defined as taking an IRCEvent, an IRCUser, an
account string, or no-param arity 0.