printer.d: Add abbreviateBadges
This takes a string of badges (from the IRCv3 badges= tag in Twitch
events) and reduces them to be represented as a string of single
characters. B for broadcaster, S for subscriber, M for moderator, etc.
This allows us to display all badges a user currently has, and not just
the most "important" one.

It doesn't solve the problem of badges bleeding into other channels.
Users can still only have one set of badges at any one time.