Add experimental Twitch awareness
Twitch does not always enumerate participants in a channel when joining
one, nor does it always echo JOINs and PARTs. Unsure why it does so on
some channels and not on others.

TwitchAwareness implements catching senders upon any simple Twitch
channel event. This is overdoing it but it may be the only way.