twitch.d: Set user class to unset when rewriting
This happens when there is an event with the server as sender but with
an obvious user origin, such as TWITCH_SUB. These are rewritten to look
like the user sent them, but information about the user class is lost in
the progress. Mark the user as class unset and make persistence.d
process it anew.