Add the notion of major and minor badges
Major badges are the kind of badges we've been dealing with so far.
Minor badges are such that relates to a specific game, always ending in
"_1". overwatch-league_1, cuphead_1, H1Z1_1, etc.

Separate the two with a colon. Still print the colon even if only a
minor badge and no major one to separate it from.