Add safeguards for plugins not getting torn down
There seems to be a race condition somewhere in which plugins don't get
properly torn down. When this happens it gets stuck at
"Exiting/Aborting...", waiting for the Pipeline plugin to exit, which it
never does. It may also be why seen users seem to randomly reset to 0.

Check if plugins have been torn down at late shutdown, and tear them
down if not. Additionally, for now check if the Pipeline's fifo still
exists and remove it if so. THIS IS NOT A GOOD SOLUTION TO THE PROBLEM.
But it's what we have to work with at this point.