printer.d: Never manually update buffer months
The `periodically` function rotates buffers once a day by committing
the lines to disk and removing the buffers, at midnight. By extension, a
new month will never start without a day first having ended.

Even in a pathological scenario where several ListeningAttempts were
queued to be parsed and acted upon, in the sense that it would take an
arbitrary amount of seconds for them all to trigger (in which time
`periodically` will not run), manually updating month and inserting a
datestamp will only make it clear the buffer and reinsert the datestamp
once `periodically` comes around.

It's not perfect and some events may technically be saved as belonging
to one month when it was actually processed at the beginning of the
next, but it's such a remote possibility we don't really care. If it
turns out we do, just move the `periodically` loop into the
listenerloop, in main.d mainLoop. There will be more overhead but it's
also miniscule.