pipeline.d: Rework handling of $TMPDIR
$TMPDIR is a Posix variable meant to override the otherwise hardcoded
temporary directory location of /tmp. The problem is that while this
should still equate to /tmp on all platforms under version(Posix), it
doesn't on OSX where its value will be a long, programmatically
generated path, making it unsuited for keeping a temporary file we want
to easily pipe text to.

Version out obeying $TMPDIR on OSX. Set version OSXTMPDIR in dub.json or
uncomment the version specifier at the top of pipeline.d to change the
behaviour.