Break IRCPluginImpl.onEvent into onEvent and onEventImpl
With an onEvent that only calls onEventImpl, we allow for plugins to
override it (as that's how mixins work) and inject their own code, while
still remaining able to continue with the function-triggering by calling
onEventImpl.

This is mostly for use with Twitch plugins and similar that only work on
a particular daemon.