Replace serverguards in automodes.d with onEvent injection
Much like in twitchbot.d we define a void onEvent(IRCEvent) in the
IRCPlugin itself, which makes the IRCPluginImpl mixin not mix in its
onEvent. The real one is onEventImpl however, and so we pass the event
onto it iff we're not connected to a Twitch server. In which case
nothing happens.