Add nom with Flag!"inherit"
At several places in the code we have a string that we don't know
whether or not it contains a separator. If it does, we want to nom it
and separately deal with the returned value and the remaining, advanced
string. But if it doesn't, we want the returned value to inherit the
passed string, and clear it afterwards.

string slice = "foo";
string verb;

if (slice.contains(' '))
{
    verb = slice.nom(' ');
}
else
{
    verb = slice;  // "foo"
    slice = string.init;
}

This adds an overload of nom that takes a flag to automate this.

string slice = "foo";
immutable verb = slice.nom!(Yes.inherit)(' ');