Private notes
Note additions sent in a private query message are inherently not
related to a channel, and will as such be treated as private notes. They
are replayed just as "channel" notes are; when a user joins a channel or
shows activity in it, except the note is replayed in a private query.