Raise Timeout.whoisRetry from 60 to 300 (seconds)
Determining a good default for this is tricky. On one hand we'll want
know when someone logs in as soon as possible, but on the other hand
that's a lot of WHOISes in a busy channel. If people don't log in, it
will endlessly WHOIS.

We get logins for free on freenode, where such is included in JOINs (and
announced when done after that with ACCOUNTs), so it's less of an issue
there.

Revisit and tweak. Should it be more or less?