Fix broken tryImportMixin
It was judging whether a plugin was available based on the availability
of the module, which makes sense if you want to disable a module by
removing its file, but not if you disable it with
version(WithSomePlugin): at the top of the file.

Make it test for __traits(compiles, __traits(identifier, module.Plugin))
instead of merely the module.