Fix connecting with a wanted nickname
We were expecting to see the effective nickname as the target nickname
in RPL_WELCOME, but that type was not being exempted from having the
nickname cleared (when it's the bot's) in postParseSanityCheck in
irc/parsing.d.

Only inherit the nickname if it's non-zero length, and exempt
RPL_WELCOME from having the target zeroed out.