Optimise isEnabled() by introspecting Settings
IRCPluginImpl.isEnabled() works by iterating all member variables in the
IRCPlugin it's mixed into, looking for a struct variable annotated Settings;
then it iterates all the members of that struct, looking for a bool
variable annotated Enabler; then it returns the value of that bool.

You could get pathological edge-cases where an exaggerated IRCPlugin with a
million member variables, where the Settings-annotated one is placed at the
very end. It would have to iterate through all variables until it got
to the correct one. Then the annotated struct could have hundreds of
variables too, with the Enabler at its very end. You could even have
cases where there are no Settings and no Enablers, in which case it
would still iterate through everything. It's not clever.

So instead, skip iterating all IRCPlugins for which we know there is no
Settings struct, and if there is, skip iterating its member variables if
we know there is no Enabler amongst them.