Use a Fiber when checking for concurrency messages
One big drawback of using concurrency messages to send information back
from plugins to the main loop is that the current design that doesn't
use globals incurs an allocation upon entering the function that
receives them. If we do it in a resumable Fiber, the function
technically only be entered once.