Default joinOnInvite to false
It was honestly mostly set to true to be able to join #d from the
overflow #d.unregistered (or similar), which you were forwarded to if
you hadn't registered on services. This doesn't seem to be the case
anymore.

Seeing as it now also doubles as following hosts on Twitch it makes
sense to have it default to disabled.