Backpedal and remove ALL Twitch raid logic
We were working under the false assumption that TWITCH_RAID fired in the
originating channel. Turns out it fires in the target one.

There doesn't seem to be a way for us to know when a raid starts and
when it stops. If we find one, revisit, because this feature seemed fun.