Add special logic for anonsubgift and anongiftpaidupgrade
Official documents say they set the sender to be the channel owner, in
which case we want to somehow want to convey that it's anonymous. For
instance by setting the sender nickname empty and the sender address to
the server's.

Still unsure about the full tags, so add some warnings to help us
(given version TwitchWarnings).