Gate all the dummy Twitch cases behind version TwitchSupport
The point of having an entry for every known tag was to be able to react
when it fell to the default case. Since a message is then only printed
if version TwitchSupport, it served no purpose on other builds.

We may as well then "hide" all the empty known tags behind version
TwitchSupport and save normal builds some cycles.