Fix potential Twitch tag race condition on anon gifts
These events are sent with a sender of ananonymousgifter (display
name/alias AnAnonymousGifter). To properly convey that the sender is
anonymous we clear it by setting the sender nickname and alias_ to empty
strings. This, however, assumes that the display name tag doesn't come
after the "subgift" etc msg-id. It breaks if it does, in the sense that
we get a sender without a nickname yet with an alias_.

Manually clear the alias_ if it exists when no nickname does.