v1.6.0
This is a minor release that adds some functionality and removes others.

Notably the ability to generate assert blocks with `--asserts` is now
removed. It has been included in upstream dialect, where it belongs.

If you run into problems compiling with dmd on Windows, try the
`--build-mode=singleFile` dub flag. It will take a long time but at
least it will compile. An alternative is to use ldc, with which
`singleFile` is not needed.

dialect, lu, requests and arsd-official dependency versions were bumped.
Remember to `dub upgrade`.

* Greatly lowered the pressure on the garbage collector by using lazy
  functions in key places (generating fewer new strings)
* Improved CI; now tests several compiler versions
* Removed debugging.d outright. Formatting deltas of structs is now in
  upstream lu:deltastrings, and generation of asserts is in dialect.
* Finally fixed Twitch no colour on first message
* Minor tweaks and fixes

Please report bugs. Feedback welcome.