v1.6.1
This is an emergency patch fixing a version name clash with dialect,
preventing compilation of unittest and dev builds.

The build command dub calls included -version=AssertGeneration, which we
used to gate assert generation in the Admin plugin. This allowed us to
only build in the functionality for appropriate build configurations.

However, when building kameloso it also specified dialect's source files
as arguments to the compiler, building them all in one go. This meant
kameloso's AssertGeneration version bled into dialect, causing it to
erroneously include the `assertgen.d` file, which should otherwise only
be built with the build configuration `assertgen`. It contains a main
function, which kameloso also provides, and which there cannot be more
than one of.

Renamed kameloso's AssertGeneration version to AdminAssertGeneration, to
avoid this collision. dialect remains unchanged.