Snapshot settings.exitSummary at start of mainLoop
If the setting is changed mid-execution (and mid-connection), the
connection start was never recorded, and so no meaningful information
can be presented in the summary. Likewise if it is disabled
mid-execution the stop time won't be recorded.

If we snapshot the setting and only obey what was set during the start
of the main loop, any such changes are safely ignored.

Subsequent connections will make new snapshots and create proper entries.