IRCPluginState as ref parameter (in more places)
It's huge so we really don't want to copy it more than is absolutely
necessary. We can't `@disable this(this)` as this breaks making easy
copies of it for concurrency messages. It's possible to create a `.dup`
member function that returns by value, but there's little gain to be had
and more complexity.

Just pass by (const) ref wherever appropriate.