Remove twitchbot.d onUserState
The purpose of the trigger was to update the list of channel operators
if we have a moderator badge, so we would know if it's safe to send
messages faster or not (moderators are rate-limited less). As it turns
out however, the main loop isn't privy to this plugin state-specific
information, so the information is lost anyway.

The current approach is to just blindly assume we're moderator in all
home channels. As such we can skip checking every single USERSTATE,
making this trigger function unnecessary.