More channel AA checks to avoid out of range access
On Twitch it seems that you can see events from channels you have yet to
SELFJOIN, if you queue to join too any at once. We relied on SELFJOIN to
set up the channel and then blindly assumed it would be there on other
events, which is true on normal IRC servers. But not Twitch.

Add simple (event.channel in plugin.state.channels) checks to ward
against this.

Also make the use of it uniformly styled.