Static checks for plugins/common.d mixin templates
We have three kinds of mixin templates; those that are to be mixed into
a module-level scope, those to be mixed into an IRCPlugin subclass, and
those to be mixed into a function.

Add static asserts to stop compilation if they're mixed into the wrong
context.

There's a whole bunch of copy/paste going on currently, so future
improvements will include some way to deduplicate the checks.