Only use __traits(isModule) if __VERSION__ >= 2087
The trait seems to have been added with 2.087. Sadly this means that for
compilers below that version we lose out on the static constraints for
module-level template mixins completely, and we get a less-helpful
error message for the other ones when wrongly mixed into a module scope.

But with 2.091 released that's mostly academic at this point.