Fix a rare case of unreachable code in onEventImpl
We were returning from inside a foreach loop which had an additional
return statement after it, so the compiler could sometimes statically
infer that the outer return would never be called, and as such fail
compilation.

Now we set a return value and break out of the loop, only using that one
particular exit point.