dub.sdl 2.0
Not as drastic a change as in lu and dialect, but still considerable.

The pluginless configurations are gone. As is configurations `posix` and
`windows`, which are now baked into a universal `application`. It only
made sense at a time when we couldn't do colours on Windows.

Made versions lists vertical.

Updated arsd-official dependency version to 7.1.1..