v1.8.0
This is a minor update but with lots of large changes. Some of it was
exported to lu before making it to release, however. Other bits were
imported from it, judged to not be generic enough.

Several new modules were added to combat "common bloat", or how our
`common.d` files were growing increasingly large in both size and scope.

Most plugins should not need to be changed in any way. One exception
would be if they had final switches over `IRCUser.Class`, or if they
were annotated with a `BotRegex` that took a `Regex` parameter (as
opposed to string regex expression).

Other changes:

* Updated lu and dialect dependencies
* Added module `kameloso.config`
* Added module `kameloso.platform`
* Added module `kameloso.semver`
* Added module `kameloso.plugins.ircplugin`
* Added module `kameloso.plugins.awareness`
* Mixins are now statically constrained to a specific scope
* `BotRegex` improved. Now no longer takes a `Regex` parameter.
* `BotRegex`es are now included in the Help listing
* Changes to interface `IRCPlugin`; some new things added, some things
  removed
* dialect removed the notion of `IRCUser.Class.special`, so so did we
* Replaced `Labeled!(Fiber, long)` with `Tuple!(Fiber, long)`
* Centralised tinting into one struct with a static opDispatch, making
  for less boilerplate
* BashQuote plugin merged with the Chatbot plugin
* Slimmed `kameloso.common` down by moving things into other modules
* Likewise slimmed `kameloso.plugins.common` down
* Revamped the Admin and Notes plugins
* Better error messages in the main loop
* Much more

Feedback appreciated.